Handle document retrieval when page being referenced is not in the docs directory
We fallback to the project's readme (with a myriad of extensions, but just one casing) when the docs directory doesn't have an `index.md` file.

We also cleverly replace any urls in said file to remove the `docs/` prefix, which allows readme writers to assume it will work properly on github *and* viewdocs

Finally, we also allow fallbacks for any files *not* found in the docs dir. This helps cases where the readme may link to a `CONTRIBUTING.md` or similar.

Closes #33