Properly handle linking to versioned docs
Previously, you could externally link to a versioned copy of a given doc
page, but navigating with that version was not properly supported. This made
it difficult to provide support for major version changes within dokku.

This change does two things:

- Strips the hostname from the link
- Replaces bare repository link prefixes with a versioned link

With the former change, the following link:

    <a href="http://progrium.dokku.io/viewdocs/development">Development</a>

Becomes:

    <a href="/viewdocs/development">Development</a>

This makes it easy to do the second change (assuming the branch name `dev`):

    <a href="/viewdocs/development">Development</a>

Becomes:

    <a href="/viewdocs~dev/development">Development</a>